LeastSquaresOptim.jl

This is an extension for importing solvers from LeastSquaresOptim.jl into the SciML interface. Note that these solvers do not come by default, and thus one needs to install the package before using these solvers:

using Pkg
Pkg.add("LeastSquaresOptim")
using LeastSquaresOptim, NonlinearSolve

Solver API

NonlinearSolve.LeastSquaresOptimJL — Type

LeastSquaresOptimJL(alg = :lm; linsolve = nothing, autodiff::Symbol = :central)

Wrapper over LeastSquaresOptim.jl for solving NonlinearLeastSquaresProblem.

Arguments

alg: Algorithm to use. Can be :lm or :dogleg.

Keyword Arguments

linsolve: Linear solver to use. Can be :qr, :cholesky or :lsmr. If nothing, then LeastSquaresOptim.jl will choose the best linear solver based on the Jacobian structure.

autodiff: Automatic differentiation / Finite Differences. Can be :central or :forward.

Note

This algorithm is only available if LeastSquaresOptim.jl is installed.