SpeedMapping.jl

This is a extension for importing solvers from SpeedMapping.jl into the SciML interface. Note that these solvers do not come by default, and thus one needs to install the package before using these solvers:

using Pkg
Pkg.add("SpeedMapping")
using SpeedMapping, NonlinearSolve

Solver API

NonlinearSolve.SpeedMappingJL — Type

SpeedMappingJL(; σ_min = 0.0, stabilize::Bool = false, check_obj::Bool = false,
    orders::Vector{Int} = [3, 3, 2], time_limit::Real = 1000)

Wrapper over SpeedMapping.jl for solving Fixed Point Problems. We allow using this algorithm to solve root finding problems as well.

Keyword Arguments

σ_min: Setting to 1 may avoid stalling (see [13]).

stabilize: performs a stabilization mapping before extrapolating. Setting to true may improve the performance for applications like accelerating the EM or MM algorithms (see [13]).

check_obj: In case of NaN or Inf values, the algorithm restarts at the best past iterate.

orders: determines ACX's alternating order. Must be between 1 and 3 (where 1 means no extrapolation). The two recommended orders are [3, 2] and [3, 3, 2], the latter being potentially better for highly non-linear applications (see [13]).

time_limit: time limit for the algorithm.

Note

This algorithm is only available if SpeedMapping.jl is installed.