Logging the Solve Process

All NonlinearSolve.jl native solvers allow storing and displaying the trace of the nonlinear solve process. This is controlled by 3 keyword arguments to solve:

show_trace: Must be Val(true) or Val(false). This controls whether the trace is displayed to the console. (Defaults to Val(false))

trace_level: Needs to be one of Trace Objects: TraceMinimal, TraceWithJacobianConditionNumber, or TraceAll. This controls the level of detail of the trace. (Defaults to TraceMinimal())

store_trace: Must be Val(true) or Val(false). This controls whether the trace is stored in the solution object. (Defaults to Val(false))

Example Usage

using ModelingToolkit, NonlinearSolve

@variables x y z
@parameters σ ρ β

# Define a nonlinear system
eqs = [0 ~ σ * (y - x),
    0 ~ x * (ρ - z) - y,
    0 ~ x * y - β * z]
@named ns = NonlinearSystem(eqs, [x, y, z], [σ, ρ, β])

u0 = [x => 1.0, y => 0.0, z => 0.0]

ps = [σ => 10.0 ρ => 26.0 β => 8 / 3]

prob = NonlinearProblem(ns, u0, ps)

solve(prob)

retcode: Success
u: 3-element Vector{Float64}:
 -1.5500210596853105e-15
 -1.7736939928072096e-15
  5.344012391148554e-16

This produced the output, but it is hard to diagnose what is going on. We can turn on the trace to see what is happening:

solve(prob; show_trace = Val(true), trace_level = TraceAll(10))

retcode: Success
u: 3-element Vector{Float64}:
 -1.5500210596853105e-15
 -1.7736939928072096e-15
  5.344012391148554e-16

You can also store the trace in the solution object:

sol = solve(prob; trace_level = TraceAll(), store_trace = Val(true));

sol.trace

----     -------------        -----------          -------             
Iter     f(u) inf-norm        Step 2-norm          cond(J)             
----     -------------        -----------          -------             
0        2.60000000e+01       4.94065646e-324      Inf                 
1        3.11334008e+01       5.00000000e-01       1.00000000e+00      
2        1.70049757e+01       2.47663402e+00       4.01371531e+00      
3        1.25070298e+01       4.55718519e-01       3.44264376e+00      
4        4.08844590e-02       1.26766212e+00       1.04551194e+01      
5        4.28684266e-02       1.01939613e-03       1.04494280e+01      
6        4.69551736e-02       1.54013596e-03       1.23850377e+01      
7        2.18953882e-02       9.39808481e-03       4.41202867e+01      
8        3.44645828e-03       6.91851172e-03       7.57097566e+01      
9        1.70717053e-06       1.29181254e-03       8.96736841e+01      
10       5.34816923e-08       6.20165360e-07       8.97163144e+01      
11       3.56364608e-09       1.87220799e-08       9.23959279e+01      
12       5.75065840e-10       1.55202581e-09       9.99847641e+01      
13       2.10857475e-09       5.75259279e-10       1.00000000e+00      
14       7.73144074e-09       2.15118075e-09       1.00000000e+00      
15       1.61419378e-07       1.10214442e-08       1.00000000e+00      
16       3.35883180e-06       2.01233323e-07       1.00000000e+00      
17       4.22320222e-07       2.13153891e-07       2.13128522e+01      
18       3.25869912e-06       1.98291746e-07       2.44319765e+01      
19       1.61167752e-06       2.27561826e-07       3.80693465e+01      
20       6.53190948e-06       9.59954095e-07       6.91266544e+01      
21       3.85268536e-14       7.69971908e-07       1.47593690e+01      

Note

For iteration == 0 only the norm(fu, Inf) is guaranteed to be meaningful. The other values being meaningful are solver dependent.

API

NonlinearSolve.TraceMinimal — Type

TraceMinimal(freq)
TraceMinimal(; print_frequency = 1, store_frequency::Int = 1)

Trace Minimal Information

Iteration Number

f(u) inf-norm

Step 2-norm

Arguments

freq: Sets both print_frequency and store_frequency to freq.

Keyword Arguments

print_frequency: Print the trace every print_frequency iterations if show_trace == Val(true).

store_frequency: Store the trace every store_frequency iterations if store_trace == Val(true).

NonlinearSolve.TraceWithJacobianConditionNumber — Type

TraceWithJacobianConditionNumber(freq)
TraceWithJacobianConditionNumber(; print_frequency = 1, store_frequency::Int = 1)

TraceMinimal + Print the Condition Number of the Jacobian.

Arguments

freq: Sets both print_frequency and store_frequency to freq.

Keyword Arguments

print_frequency: Print the trace every print_frequency iterations if show_trace == Val(true).

store_frequency: Store the trace every store_frequency iterations if store_trace == Val(true).

NonlinearSolve.TraceAll — Type

TraceAll(freq)
TraceAll(; print_frequency = 1, store_frequency::Int = 1)

TraceWithJacobianConditionNumber + Store the Jacobian, u, f(u), and δu.

Warning

This is very expensive and makes copyies of the Jacobian, u, f(u), and δu.

Arguments

freq: Sets both print_frequency and store_frequency to freq.

Keyword Arguments

print_frequency: Print the trace every print_frequency iterations if show_trace == Val(true).

store_frequency: Store the trace every store_frequency iterations if store_trace == Val(true).