Nonlinear Problems

The Two Types of Nonlinear Problems

NonlinearSolve.jl tackles two related types of nonlinear systems:

Interval rootfinding problems. I.e., find the $t in [t_0, t_f]$ such that f(t) = 0.

Systems of nonlinear equations, i.e. find the u such that f(u) = 0.

The former is for solving scalar rootfinding problems, i.e. finding a single number, and requires that a bracketing interval is known. For a bracketing interval, one must have that the sign of f(t_0) is opposite the sign of f(t_f), thus guaranteeing a root in the interval.

The latter type of nonlinear system can be multidimensional and thus no ordering nor boundaries are assumed to be known. For a system of nonlinear equations, f can return an array and the solver seeks to find the value of u for which all outputs of f are simultaniously zero.

Note

Interval rootfinding problems allow for f to return an array, in which case the interval rootfinding problem is interpreted as finding the first t such that any of the components of the array hit zero.

Problem Construction Details

SciMLBase.IntervalNonlinearProblem — Type

Defines a interval nonlinear system problem. Documentation Page: https://docs.sciml.ai/NonlinearSolve/stable/basics/NonlinearProblem/

Mathematical Specification of a Interval Nonlinear Problem

To define a Nonlinear Problem, you simply need to give the function $f$ which defines the nonlinear system:

\[f(t,p) = u = 0\]

along with an interval tspan is t \in [t_0,t_f] within which the root should be found. f should be specified as f(t,p) (or in-place as f(u,t,p)), and tspan should be a Tuple{T,T} where T <: Number.

Note

The output value u is not required to be a scalar. When u is an AbstractArray, the problem is a simultanious interval nonlinear problem where the solvers are made to give the first t for which any of the u hit zero. Currently none of the solvers support this mode.

Problem Type

Constructors

IntervalNonlinearProblem(f::NonlinearFunction,tspan,p=NullParameters();kwargs...)
IntervalNonlinearProblem{isinplace}(f,tspan,p=NullParameters();kwargs...)

isinplace optionally sets whether the function is in-place or not. This is determined automatically, but not inferred.

Parameters are optional, and if not given, then a NullParameters() singleton will be used, which will throw nice errors if you try to index non-existent parameters. Any extra keyword arguments are passed on to the solvers. For example, if you set a callback in the problem, then that callback will be added in every solve call.

Fields

f: The function in the problem.

tspan: The interval in which the root is to be found.

p: The parameters for the problem. Defaults to NullParameters.

kwargs: The keyword arguments passed on to the solvers.

SciMLBase.NonlinearProblem — Type

Defines a nonlinear system problem. Documentation Page: https://docs.sciml.ai/NonlinearSolve/stable/basics/NonlinearProblem/

Mathematical Specification of a Nonlinear Problem

To define a Nonlinear Problem, you simply need to give the function $f$ which defines the nonlinear system:

\[f(u,p) = 0\]

and an initial guess $u₀$ of where f(u,p)=0. f should be specified as f(u,p) (or in-place as f(du,u,p)), and u₀ should be an AbstractArray (or number) whose geometry matches the desired geometry of u. Note that we are not limited to numbers or vectors for u₀; one is allowed to provide u₀ as arbitrary matrices / higher-dimension tensors as well.

Problem Type

Constructors

NonlinearProblem(f::NonlinearFunction,u0,p=NullParameters();kwargs...)
NonlinearProblem{isinplace}(f,u0,p=NullParameters();kwargs...)

isinplace optionally sets whether the function is in-place or not. This is determined automatically, but not inferred.

Parameters are optional, and if not given, then a NullParameters() singleton will be used, which will throw nice errors if you try to index non-existent parameters. Any extra keyword arguments are passed on to the solvers. For example, if you set a callback in the problem, then that callback will be added in every solve call.

For specifying Jacobians and mass matrices, see the NonlinearFunctions page.

Fields

f: The function in the problem.

u0: The initial guess for the steady state.

p: The parameters for the problem. Defaults to NullParameters.

kwargs: The keyword arguments passed on to the solvers.