Nonlinear Solutions

SciMLBase.NonlinearSolution — Type

struct NonlinearSolution{T, N, uType, R, P, A, O, uType2, S, Tr} <: SciMLBase.AbstractNonlinearSolution{T, N}

Representation of the solution to a nonlinear equation defined by a NonlinearProblem, or the steady state solution to a differential equation defined by a SteadyStateProblem.

Fields

u: the representation of the nonlinear equation's solution.

resid: the residual of the solution.

prob: the original NonlinearProblem/SteadyStateProblem that was solved.

alg: the algorithm type used by the solver.

original: if the solver is wrapped from an alternative solver ecosystem, such as NLsolve.jl, then this is the original return from said solver library.

retcode: the return code from the solver. Used to determine whether the solver solved successfully or whether it exited due to an error. For more details, see the return code documentation.

left: if the solver is bracketing method, this is the final left bracket value.

right: if the solver is bracketing method, this is the final right bracket value.

stats: statistics of the solver, such as the number of function evaluations required.

Return Code

ReturnCode.Success - The nonlinear solve succeeded.

ReturnCode.ConvergenceFailure - The nonlinear solve failed to converge due to stalling or some limit of the solver was exceeded. For example, too many shrinks for trust region methods, number of resets for Broyden, etc.

ReturnCode.Unstable - This corresponds to NonlinearSafeTerminationReturnCode.ProtectiveTermination and is caused if the step-size of the solver was too large or the objective value became non-finite.

ReturnCode.MaxIters - The maximum number of iterations was reached.

ReturnCode.Failure - The nonlinear solve failed for some reason. This is used sparingly and mostly for wrapped solvers for which we don't have a better error code.