Termination Conditions

Provides a API to specify termination conditions for NonlinearProblem and SteadyStateProblem. For details on the various termination modes:

Termination Conditions

The termination condition is constructed as:

cache = init(du, u, AbsNormTerminationMode(); abstol = 1e-9, reltol = 1e-9)

If abstol and reltol are not supplied, then we choose a default based on the element types of du and u.

We can query the cache using DiffEqBase.get_termination_mode, DiffEqBase.get_abstol and DiffEqBase.get_reltol.

To test for termination simply call the cache:

terminated = cache(du, u, uprev)

Note

The default for NonlinearSolve.jl is AbsSafeBestTerminationMode!

Absolute Tolerance

DiffEqBase.AbsTerminationMode — Type

AbsTerminationMode <: AbstractNonlinearTerminationMode

Terminates if $all \left( | \frac{\partial u}{\partial t} | \leq abstol \right)$.

DiffEqBase.AbsNormTerminationMode — Type

AbsNormTerminationMode <: AbstractNonlinearTerminationMode

Terminates if $\| \frac{\partial u}{\partial t} \| \leq abstol$.

DiffEqBase.AbsSafeTerminationMode — Type

AbsSafeTerminationMode <: AbstractSafeNonlinearTerminationMode

Essentially AbsNormTerminationMode + terminate if there has been no improvement for the last patience_steps + terminate if the solution blows up (diverges).

Constructor

AbsSafeTerminationMode(; protective_threshold = 1e3, patience_steps = 100,
                        patience_objective_multiplier = 3, min_max_factor = 1.3)

DiffEqBase.AbsSafeBestTerminationMode — Type

AbsSafeBestTerminationMode <: AbstractSafeBestNonlinearTerminationMode

Essentially AbsSafeTerminationMode, but caches the best solution found so far.

Constructor

AbsSafeBestTerminationMode(; protective_threshold = 1e3, patience_steps = 100,
                            patience_objective_multiplier = 3, min_max_factor = 1.3)

Relative Tolerance

DiffEqBase.RelTerminationMode — Type

RelTerminationMode <: AbstractNonlinearTerminationMode

Terminates if $all \left(| \frac{\partial u}{\partial t} | \leq reltol \times | u | \right)$.

DiffEqBase.RelNormTerminationMode — Type

RelNormTerminationMode <: AbstractNonlinearTerminationMode

Terminates if $\| \frac{\partial u}{\partial t} \| \leq reltol \times \| \frac{\partial u}{\partial t} + u \|$

DiffEqBase.RelSafeTerminationMode — Type

RelSafeTerminationMode <: AbstractSafeNonlinearTerminationMode

Essentially RelNormTerminationMode + terminate if there has been no improvement for the last patience_steps + terminate if the solution blows up (diverges).

Constructor

RelSafeTerminationMode(; protective_threshold = 1e3, patience_steps = 100,
                         patience_objective_multiplier = 3, min_max_factor = 1.3)

DiffEqBase.RelSafeBestTerminationMode — Type

RelSafeBestTerminationMode <: AbstractSafeBestNonlinearTerminationMode

Essentially RelSafeTerminationMode, but caches the best solution found so far.

Constructor

RelSafeBestTerminationMode(; protective_threshold = 1e3, patience_steps = 100,
                            patience_objective_multiplier = 3, min_max_factor = 1.3)

Both Absolute and Relative Tolerance

DiffEqBase.NormTerminationMode — Type

NormTerminationMode <: AbstractNonlinearTerminationMode

Terminates if $\| \frac{\partial u}{\partial t} \| \leq reltol \times \| \frac{\partial u}{\partial t} + u \|$ or $\| \frac{\partial u}{\partial t} \| \leq abstol$

DiffEqBase.SteadyStateDiffEqTerminationMode — Type

SteadyStateDiffEqTerminationMode <: AbstractNonlinearTerminationMode

Check if all values of the derivative is close to zero wrt both relative and absolute tolerance.

The default used in SteadyStateDiffEq.jl! Not recommended for large problems, since the convergence criteria is very strict and never reliably satisfied for most problems.

DiffEqBase.SimpleNonlinearSolveTerminationMode — Type

SimpleNonlinearSolveTerminationMode <: AbstractNonlinearTerminationMode

Check if all values of the derivative is close to zero wrt both relative and absolute tolerance. Or check that the value of the current and previous state is within the specified tolerances.

The default used in SimpleNonlinearSolve.jl! Not recommended for large problems, since the convergence criteria is very strict and never reliably satisfied for most problems.

Return Codes

DiffEqBase.NonlinearSafeTerminationReturnCode — Module

NonlinearSafeTerminationReturnCode

Return Codes for the safe nonlinear termination conditions.

DiffEqBase.NonlinearSafeTerminationReturnCode.Success — Constant

NonlinearSafeTerminationReturnCode.Success

Termination Condition was satisfied!

DiffEqBase.NonlinearSafeTerminationReturnCode.Default — Constant

NonlinearSafeTerminationReturnCode.Default

Default Return Code. Used for type stability and conveys no additional information!

DiffEqBase.NonlinearSafeTerminationReturnCode.Failure — Constant

NonlinearSafeTerminationReturnCode.Failure

Termination Condition was not satisfied!

DiffEqBase.NonlinearSafeTerminationReturnCode.PatienceTermination — Constant

NonlinearSafeTerminationReturnCode.PatienceTermination

Terminate if there has been no improvement for the last patience_steps.

DiffEqBase.NonlinearSafeTerminationReturnCode.ProtectiveTermination — Constant

NonlinearSafeTerminationReturnCode.ProtectiveTermination

Terminate if the objective value increased by this factor wrt initial objective or the value diverged.

[Deprecated] Termination Condition API

Warning

This is deprecated. Currently only parts of SimpleNonlinearSolve uses this API. That will also be phased out soon!

DiffEqBase.NLSolveTerminationCondition — Type

NLSolveTerminationCondition(mode; abstol::T = 1e-8, reltol::T = 1e-6,
                            protective_threshold = 1e3, patience_steps::Int = 30,
                            patience_objective_multiplier = 3, min_max_factor = 1.3)

Define the termination criteria for the NonlinearProblem or SteadyStateProblem.

Termination Conditions

Termination on Absolute Tolerance

NLSolveTerminationMode.Abs: Terminates if $all \left( | \frac{\partial u}{\partial t} | \leq abstol \right)$

NLSolveTerminationMode.AbsNorm: Terminates if $\| \frac{\partial u}{\partial t} \| \leq abstol$

NLSolveTerminationMode.AbsSafe: Essentially abs_norm + terminate if there has been no improvement for the last 30 steps + terminate if the solution blows up (diverges)

NLSolveTerminationMode.AbsSafeBest: Same as NLSolveTerminationMode.AbsSafe but uses the best solution found so far, i.e. deviates only if the solution has not converged

Termination on Relative Tolerance

NLSolveTerminationMode.Rel: Terminates if $all \left(| \frac{\partial u}{\partial t} | \leq reltol \times | u | \right)$

NLSolveTerminationMode.RelNorm: Terminates if $\| \frac{\partial u}{\partial t} \| \leq reltol \times \| \frac{\partial u}{\partial t} + u \|$

NLSolveTerminationMode.RelSafe: Essentially rel_norm + terminate if there has been no improvement for the last 30 steps + terminate if the solution blows up (diverges)

NLSolveTerminationMode.RelSafeBest: Same as NLSolveTerminationMode.RelSafe but uses the best solution found so far, i.e. deviates only if the solution has not converged

Termination using both Absolute and Relative Tolerances

NLSolveTerminationMode.Norm: Terminates if $\| \frac{\partial u}{\partial t} \| \leq reltol \times \| \frac{\partial u}{\partial t} + u \|$ or $\| \frac{\partial u}{\partial t} \| \leq abstol$

NLSolveTerminationMode.SteadyStateDefault: Check if all values of the derivative is close to zero wrt both relative and absolute tolerance. This is usable for small problems but doesn't scale well for neural networks.

NLSolveTerminationMode.NLSolveDefault: Check if all values of the derivative is close to zero wrt both relative and absolute tolerance. Or check that the value of the current and previous state is within the specified tolerances. This is usable for small problems but doesn't scale well for neural networks.

General Arguments

abstol: Absolute Tolerance

reltol: Relative Tolerance

Arguments specific to *Safe* modes

protective_threshold: If the objective value increased by this factor wrt initial objective terminate immediately.

patience_steps: If objective is within patience_objective_multiplier factor of the criteria and no improvement within min_max_factor has happened then terminate.

Warning

This has been deprecated and will be removed in the next major release. Please use the new dispatch based termination conditions API.