Automatic Differentiation Backends

Summary of Finite Differencing Backends

AutoFiniteDiff: Finite differencing, not optimal but always applicable.

AutoSparseFiniteDiff: Sparse version of AutoFiniteDiff.

Summary of Forward Mode AD Backends

AutoForwardDiff: The best choice for dense problems.

AutoSparseForwardDiff: Sparse version of AutoForwardDiff.

AutoPolyesterForwardDiff: Might be faster than AutoForwardDiff for large problems. Requires PolyesterForwardDiff.jl to be installed and loaded.

Summary of Reverse Mode AD Backends

AutoZygote: The fastest choice for non-mutating array-based (BLAS) functions.

AutoSparseZygote: Sparse version of AutoZygote.

AutoEnzyme: Uses Enzyme.jl Reverse Mode and should be considered experimental.

Note

If PolyesterForwardDiff.jl is installed and loaded, then SimpleNonlinearSolve.jl will automatically use AutoPolyesterForwardDiff as the default AD backend.

Note

The Sparse versions of the methods refers to automated sparsity detection. These methods automatically discover the sparse Jacobian form from the function f. Note that all methods specialize the differentiation on a sparse Jacobian if the sparse Jacobian is given as prob.f.jac_prototype in the NonlinearFunction definition, and the AutoSparse here simply refers to whether this jac_prototype should be generated automatically. For more details, see SparseDiffTools.jl and Sparsity Detection Manual Entry.

API Reference

Finite Differencing Backends

ADTypes.AutoFiniteDiff — Type

AutoFiniteDiff(; fdtype = Val(:forward), fdjtype = fdtype, fdhtype = Val(:hcentral))

This uses FiniteDiff.jl. While not necessarily the most efficient, this is the only choice that doesn't require the f function to be automatically differentiable, which means it applies to any choice. However, because it's using finite differencing, one needs to be careful as this procedure introduces numerical error into the derivative estimates.

Compatible with GPUs

Can be used for Jacobian-Vector Products (JVPs)

Can be used for Vector-Jacobian Products (VJPs)

Supports both inplace and out-of-place functions

Keyword Arguments

fdtype: the method used for defining the gradient

fdjtype: the method used for defining the Jacobian of constraints.

fdhtype: the method used for defining the Hessian

ADTypes.AutoSparseFiniteDiff — Type

AutoSparseFiniteDiff()

Sparse Version of AutoFiniteDiff that uses FiniteDiff.jl and the column color vector of the Jacobian Matrix to efficiently compute the Sparse Jacobian.

Supports both inplace and out-of-place functions

Forward Mode AD Backends

ADTypes.AutoForwardDiff — Type

AutoForwardDiff(; chunksize = nothing, tag = nothing)
AutoForwardDiff{chunksize, tagType}(tag::tagType)

This uses the ForwardDiff.jl package. It is the fastest choice for square or wide systems. It is easy to use and compatible with most Julia functions which have loose type restrictions.

Compatible with GPUs

Can be used for Jacobian-Vector Products (JVPs)

Supports both inplace and out-of-place functions

For type-stability of internal operations, a positive chunksize must be provided.

Keyword Arguments

chunksize: Count of dual numbers that can be propagated simultaneously. Setting this number to a high value will lead to slowdowns. Use NonlinearSolve.pickchunksize to get a proper value.

tag: Used to avoid perturbation confusion. If set to nothing, we use a custom tag.

ADTypes.AutoSparseForwardDiff — Type

AutoSparseForwardDiff(; chunksize = nothing, tag = nothing)
AutoSparseForwardDiff{chunksize, tagType}(tag::tagType)

Sparse Version of AutoForwardDiff that uses ForwardDiff.jl and the column color vector of the Jacobian Matrix to efficiently compute the Sparse Jacobian.

Supports both inplace and out-of-place functions

For type-stability of internal operations, a positive chunksize must be provided.

Keyword Arguments

chunksize: Count of dual numbers that can be propagated simultaneously. Setting this number to a high value will lead to slowdowns. Use NonlinearSolve.pickchunksize to get a proper value.

tag: Used to avoid perturbation confusion. If set to nothing, we use a custom tag.

ADTypes.AutoPolyesterForwardDiff — Type

AutoPolyesterForwardDiff(; chunksize = nothing)

Uses PolyesterForwardDiff.jl to compute the jacobian. This is essentially parallelized ForwardDiff.jl.

Supports both inplace and out-of-place functions

Keyword Arguments

chunksize: Count of dual numbers that can be propagated simultaneously. Setting this number to a high value will lead to slowdowns. Use NonlinearSolve.pickchunksize to get a proper value.

Reverse Mode AD Backends

ADTypes.AutoZygote — Type

AutoZygote()

Uses Zygote.jl package. This is the staple reverse-mode AD that handles a large portion of Julia with good efficiency.

Compatible with GPUs

Can be used for Vector-Jacobian Products (VJPs)

Supports only out-of-place functions

For VJPs this is the current best choice. This is the most efficient method for long jacobians.

ADTypes.AutoSparseZygote — Type

AutoSparseZygote()

Sparse version of AutoZygote that uses Zygote.jl and the row color vector of the Jacobian Matrix to efficiently compute the Sparse Jacobian.

Supports only out-of-place functions

This is efficient only for long jacobians or if the maximum value of the row color vector is significantly lower than the maximum value of the column color vector.

ADTypes.AutoEnzyme — Type

AutoEnzyme()

Uses reverse mode Enzyme.jl. This is currently experimental, and not extensively tested on our end. We only support Jacobian construction and VJP support is currently not implemented.

Supports both inplace and out-of-place functions

SparseDiffTools.AutoSparseEnzyme — Type

AutoSparseEnzyme()

Sparse version of AutoEnzyme that uses Enzyme.jl and the row color vector of the Jacobian Matrix to efficiently compute the Sparse Jacobian.

Supports both inplace and out-of-place functions

This is efficient only for long jacobians or if the maximum value of the row color vector is significantly lower than the maximum value of the column color vector.