Termination Conditions

Provides a API to specify termination conditions for NonlinearProblem and SteadyStateProblem. For details on the various termination modes:

Termination Conditions

The termination condition is constructed as:

cache = init(du, u, AbsSafeBestTerminationMode(); abstol = 1e-9, reltol = 1e-9)

If abstol and reltol are not supplied, then we choose a default based on the element types of du and u.

We can query the cache using DiffEqBase.get_termination_mode, DiffEqBase.get_abstol and DiffEqBase.get_reltol.

To test for termination simply call the cache:

terminated = cache(du, u, uprev)

Absolute Tolerance

DiffEqBase.AbsTerminationMode — Type

AbsTerminationMode <: AbstractNonlinearTerminationMode

Terminates if $all \left( | \frac{\partial u}{\partial t} | \leq abstol \right)$. .

DiffEqBase.AbsNormTerminationMode — Type

AbsNormTerminationMode <: AbstractNonlinearTerminationMode

Terminates if $\| \frac{\partial u}{\partial t} \| \leq abstol$ .

Constructor

AbsNormTerminationMode(internalnorm = nothing)

where internalnorm is the norm to use for the termination condition. Special handling is done for norm(_, 2), norm(_, Inf), and maximum(abs, _).

Default is left as nothing, which allows upstream frameworks to choose the correct norm based on the problem type. If directly using the init API, a proper norm must be provided.

DiffEqBase.AbsSafeTerminationMode — Type

AbsSafeTerminationMode <: DiffEqBase.AbstractSafeNonlinearTerminationMode

Essentially AbsNormTerminationMode + terminate if there has been no improvement for the last patience_steps + terminate if the solution blows up (diverges).

Constructor

AbsSafeTerminationMode(internalnorm = nothing; protective_threshold = nothing,
    patience_steps = 100, patience_objective_multiplier = 3,
    min_max_factor = 1.3, max_stalled_steps = nothing)

where internalnorm is the norm to use for the termination condition. Special handling is done for norm(_, 2), norm(_, Inf), and maximum(abs, _).

Default is left as nothing, which allows upstream frameworks to choose the correct norm based on the problem type. If directly using the init API, a proper norm must be provided.

DiffEqBase.AbsSafeBestTerminationMode — Type

AbsSafeBestTerminationMode <: DiffEqBase.AbstractSafeBestNonlinearTerminationMode

Essentially AbsSafeTerminationMode, but caches the bestsolution found so far.

Constructor

AbsSafeBestTerminationMode(internalnorm = nothing; protective_threshold = nothing,
    patience_steps = 100, patience_objective_multiplier = 3,
    min_max_factor = 1.3, max_stalled_steps = nothing)

where internalnorm is the norm to use for the termination condition. Special handling is done for norm(_, 2), norm(_, Inf), and maximum(abs, _).

Default is left as nothing, which allows upstream frameworks to choose the correct norm based on the problem type. If directly using the init API, a proper norm must be provided.

Relative Tolerance

DiffEqBase.RelTerminationMode — Type

RelTerminationMode <: AbstractNonlinearTerminationMode

Terminates if $all \left(| \frac{\partial u}{\partial t} | \leq reltol \times | u | \right)$. .

DiffEqBase.RelNormTerminationMode — Type

RelNormTerminationMode <: AbstractNonlinearTerminationMode

Terminates if $\| \frac{\partial u}{\partial t} \| \leq reltol \times \| \frac{\partial u}{\partial t} + u \|$ .

Constructor

RelNormTerminationMode(internalnorm = nothing)

where internalnorm is the norm to use for the termination condition. Special handling is done for norm(_, 2), norm(_, Inf), and maximum(abs, _).

Default is left as nothing, which allows upstream frameworks to choose the correct norm based on the problem type. If directly using the init API, a proper norm must be provided.

DiffEqBase.RelSafeTerminationMode — Type

RelSafeTerminationMode <: DiffEqBase.AbstractSafeNonlinearTerminationMode

Essentially RelNormTerminationMode + terminate if there has been no improvement for the last patience_steps + terminate if the solution blows up (diverges).

Constructor

RelSafeTerminationMode(internalnorm = nothing; protective_threshold = nothing,
    patience_steps = 100, patience_objective_multiplier = 3,
    min_max_factor = 1.3, max_stalled_steps = nothing)

where internalnorm is the norm to use for the termination condition. Special handling is done for norm(_, 2), norm(_, Inf), and maximum(abs, _).

Default is left as nothing, which allows upstream frameworks to choose the correct norm based on the problem type. If directly using the init API, a proper norm must be provided.

DiffEqBase.RelSafeBestTerminationMode — Type

RelSafeBestTerminationMode <: DiffEqBase.AbstractSafeBestNonlinearTerminationMode

Essentially RelSafeTerminationMode, but caches the bestsolution found so far.

Constructor

RelSafeBestTerminationMode(internalnorm = nothing; protective_threshold = nothing,
    patience_steps = 100, patience_objective_multiplier = 3,
    min_max_factor = 1.3, max_stalled_steps = nothing)

where internalnorm is the norm to use for the termination condition. Special handling is done for norm(_, 2), norm(_, Inf), and maximum(abs, _).

Default is left as nothing, which allows upstream frameworks to choose the correct norm based on the problem type. If directly using the init API, a proper norm must be provided.

Both Absolute and Relative Tolerance

DiffEqBase.NormTerminationMode — Type

NormTerminationMode <: AbstractNonlinearTerminationMode

Terminates if $\| \frac{\partial u}{\partial t} \| \leq reltol \times \| \frac{\partial u}{\partial t} + u \|$ or $\| \frac{\partial u}{\partial t} \| \leq abstol$ .

Constructor

NormTerminationMode(internalnorm = nothing)

where internalnorm is the norm to use for the termination condition. Special handling is done for norm(_, 2), norm(_, Inf), and maximum(abs, _).

Default is left as nothing, which allows upstream frameworks to choose the correct norm based on the problem type. If directly using the init API, a proper norm must be provided.

DiffEqBase.SteadyStateDiffEqTerminationMode — Type

SteadyStateDiffEqTerminationMode <: AbstractNonlinearTerminationMode

Check if all values of the derivative is close to zero wrt both relative and absolute tolerance.

Danger

This has been deprecated.

The following was named to match an older version of SimpleNonlinearSolve. It is currently not used as a default anywhere.

DiffEqBase.SimpleNonlinearSolveTerminationMode — Type

SimpleNonlinearSolveTerminationMode <: AbstractNonlinearTerminationMode

Check if all values of the derivative is close to zero wrt both relative and absolute tolerance. Or check that the value of the current and previous state is within the specified tolerances.

Danger

This has been deprecated.

Return Codes (Deprecated)

These are deprecated and will be removed in a future release. Use the use_deprecated_retcodes = Val(false) option to SciMLBase.init to use the new return ReturnCode versions.

DiffEqBase.NonlinearSafeTerminationReturnCode — Module

NonlinearSafeTerminationReturnCode

Return Codes for the safe nonlinear termination conditions.

These return codes have been deprecated. Termination Conditions will return SciMLBase.Retcode.T starting from v7.

DiffEqBase.NonlinearSafeTerminationReturnCode.Success — Constant

NonlinearSafeTerminationReturnCode.Success

Termination Condition was satisfied!

DiffEqBase.NonlinearSafeTerminationReturnCode.Default — Constant

NonlinearSafeTerminationReturnCode.Default

Default Return Code. Used for type stability and conveys no additional information!

DiffEqBase.NonlinearSafeTerminationReturnCode.Failure — Constant

NonlinearSafeTerminationReturnCode.Failure

Termination Condition was not satisfied!

DiffEqBase.NonlinearSafeTerminationReturnCode.PatienceTermination — Constant

NonlinearSafeTerminationReturnCode.PatienceTermination

Terminate if there has been no improvement for the last patience_steps.

DiffEqBase.NonlinearSafeTerminationReturnCode.ProtectiveTermination — Constant

NonlinearSafeTerminationReturnCode.ProtectiveTermination

Terminate if the objective value increased by this factor wrt initial objective or the value diverged.