Line Search

A convenience wrapper over LineSearches.jl and some native Line Search methods, powered internally with fast automatic differentiation.

NonlinearSolve.LineSearch — Type

LineSearch(; method = nothing, autodiff = nothing, alpha = true)

Wrapper over algorithms from LineSearches.jl. Allows automatic construction of the objective functions for the line search algorithms utilizing automatic differentiation for fast Vector Jacobian Products.

Arguments

method: the line search algorithm to use. Defaults to nothing, which means that the step size is fixed to the value of alpha.

autodiff: the automatic differentiation backend to use for the line search. Defaults to AutoFiniteDiff(), which means that finite differencing is used to compute the VJP. AutoZygote() will be faster in most cases, but it requires Zygote.jl to be manually installed and loaded.

alpha: the initial step size to use. Defaults to true (which is equivalent to 1).

Native Line Search Methods

NonlinearSolve.LiFukushimaLineSearch — Type

LiFukushimaLineSearch(; lambda_0 = 1.0, beta = 0.5, sigma_1 = 0.001,
    eta = 0.1, nan_max_iter = 5, maxiters = 50)

A derivative-free line search and global convergence of Broyden-like method for nonlinear equations by Dong-Hui Li & Masao Fukushima. For more details see https://doi.org/10.1080/10556780008805782