SimpleNonlinearSolve.jl

These methods can be used independently of the rest of NonlinearSolve.jl

Solver API

SimpleNonlinearSolve.Bisection — Type

Bisection(; exact_left = false, exact_right = false)

A common bisection method.

Keyword Arguments

exact_left: whether to enforce whether the left side of the interval must be exactly zero for the returned result. Defaults to false.

exact_right: whether to enforce whether the right side of the interval must be exactly zero for the returned result. Defaults to false.

SimpleNonlinearSolve.Falsi — Type

Falsi: A non-allocating regula falsi method

SimpleNonlinearSolve.SimpleNewtonRaphson — Type

SimpleNewtonRaphson(; chunk_size = Val{0}(), autodiff = Val{true}(),
                                 diff_type = Val{:forward})

A low-overhead implementation of Newton-Raphson. This method is non-allocating on scalar and static array problems.

Note

As part of the decreased overhead, this method omits some of the higher level error catching of the other methods. Thus, to see better error messages, use one of the other methods like NewtonRaphson

Keyword Arguments

chunk_size: the chunk size used by the internal ForwardDiff.jl automatic differentiation system. This allows for multiple derivative columns to be computed simultaneously, improving performance. Defaults to 0, which is equivalent to using ForwardDiff.jl's default chunk size mechanism. For more details, see the documentation for ForwardDiff.jl.

autodiff: whether to use forward-mode automatic differentiation for the Jacobian. Note that this argument is ignored if an analytical Jacobian is passed; as that will be used instead. Defaults to Val{true}, which means ForwardDiff.jl is used by default. If Val{false}, then FiniteDiff.jl is used for finite differencing.

diff_type: the type of finite differencing used if autodiff = false. Defaults to Val{:forward} for forward finite differences. For more details on the choices, see the FiniteDiff.jl documentation.

SimpleNonlinearSolve.Broyden — Type

Broyden()

A low-overhead implementation of Broyden. This method is non-allocating on scalar and static array problems.

SimpleNonlinearSolve.Klement — Type

Klement()

A low-overhead implementation of Klement. This method is non-allocating on scalar problems.

Missing docstring.

Missing docstring for TrustRegion. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Ridder. Check Documenter's build log for details.