Termination Conditions

Provides a API to specify termination conditions for NonLinearProblem and SteadyStateProblem. For details on the various termination modes, i.e., NLSolveTerminationMode, see the documentation for NLSolveTerminationCondition.

Termination Condition API

DiffEqBase.NLSolveTerminationCondition — Type

NLSolveTerminationCondition(mode; abstol::T = 1e-8, reltol::T = 1e-6,
                            protective_threshold = 1e3, patience_steps::Int = 30,
                            patience_objective_multiplier = 3, min_max_factor = 1.3)

Define the termination criteria for the NonlinearProblem or SteadyStateProblem.

Termination Conditions

Termination on Absolute Tolerance

NLSolveTerminationMode.Abs: Terminates if $all \left( | \frac{\partial u}{\partial t} | \leq abstol \right)$

NLSolveTerminationMode.AbsNorm: Terminates if $\| \frac{\partial u}{\partial t} \| \leq abstol$

NLSolveTerminationMode.AbsSafe: Essentially abs_norm + terminate if there has been no improvement for the last 30 steps + terminate if the solution blows up (diverges)

NLSolveTerminationMode.AbsSafeBest: Same as NLSolveTerminationMode.AbsSafe but uses the best solution found so far, i.e. deviates only if the solution has not converged

Termination on Relative Tolerance

NLSolveTerminationMode.Rel: Terminates if $all \left(| \frac{\partial u}{\partial t} | \leq reltol \times | u | \right)$

NLSolveTerminationMode.RelNorm: Terminates if $\| \frac{\partial u}{\partial t} \| \leq reltol \times \| \frac{\partial u}{\partial t} + u \|$

NLSolveTerminationMode.RelSafe: Essentially rel_norm + terminate if there has been no improvement for the last 30 steps + terminate if the solution blows up (diverges)

NLSolveTerminationMode.RelSafeBest: Same as NLSolveTerminationMode.RelSafe but uses the best solution found so far, i.e. deviates only if the solution has not converged

Termination using both Absolute and Relative Tolerances

NLSolveTerminationMode.Norm: Terminates if $\| \frac{\partial u}{\partial t} \| \leq reltol \times \| \frac{\partial u}{\partial t} + u \|$ or $\| \frac{\partial u}{\partial t} \| \leq abstol$

NLSolveTerminationMode.SteadyStateDefault: Check if all values of the derivative is close to zero wrt both relative and absolute tolerance. This is usable for small problems but doesn't scale well for neural networks.

NLSolveTerminationMode.NLSolveDefault: Check if all values of the derivative is close to zero wrt both relative and absolute tolerance. Or check that the value of the current and previous state is within the specified tolerances. This is usable for small problems but doesn't scale well for neural networks.

General Arguments

abstol: Absolute Tolerance

reltol: Relative Tolerance

Arguments specific to *Safe* modes

protective_threshold: If the objective value increased by this factor wrt initial objective terminate immediately.

patience_steps: If objective is within patience_objective_multiplier factor of the criteria and no improvement within min_max_factor has happened then terminate.